refactor: standardize feature organization by extracting inline features
Extract all TCA features from view files to App/Features/ directory:
- RootFeature from RootView.swift (350 lines)
- LoginFeature from LoginView.swift (160 lines)
- DefaultFilesFeature from DefaultFilesView.swift (170 lines)

Update CLAUDE.md to remove outdated MVVM migration references and
document the current codebase structure with organization conventions.
